--- .postcssrc
@@ -1,6 +1,0 @@
-{
-    "plugins": {
-        "tailwindcss": {},
-        "autoprefixer": {},
-    }
-}
