+++ src/components/Navbar/buttons.tsx
@@ -1,0 +1,5 @@
+import { FC } from "react";
+
+export const NavbarButtons: FC = () => {
+    return <></>;
+};
