+++ src/components/Navbar/links.ts
@@ -1,0 +1,13 @@
+export const links: {
+    name: "Home" | "Away";
+    path: string;
+}[] = [
+    {
+        name: "Home",
+        path: "/",
+    },
+    {
+        name: "Away",
+        path: "/",
+    },
+];
